spike: async rendering
we have a lot of entrypoints/css bundles in our app, and in adding them
all I realized the build was getting a lot slower and, suspiciously, all
the sass builds took the exact same amount of time to build. Seemed like
a symptom using sync interfaces could cause, so I thought I'd try seeing
if async helped.

It did not! [Dart's docs][0] do say upfront that the async interface can
be ~2x slower, but I hoped that there'd still be a win on overall
throughput even if individual bundles were slower. For some bundles that
was the case: previously every css bundle was about ~8.6s, some smaller
ones are now ~500ms. However, the slow ones got a lot more than 2x
slower: a few are now ~18.5s, which is longer than the previous entire
build (including JS), so overall wall-clock time of my entire build went
up from ~11s to ~20s.

So this went nowhere. It's possible there's still some synchronous
chokepoints in the plugin I didn't see, but I'm out of time on this
right now. If this had been more promising I would have put in more work
to make it an option for the plugin. Committing it anyway as a
historical artifact. libsass's poor async performance seems like it must
be a design issue of some kind, so maybe they'll improve that in the
future.

[0]: https://sass-lang.com/documentation/js-api/modules#compileStringAsync