Switch to sass-embedded
Hunting down some more reading on sass perf, I found this gh issue:
https://github.com/sass/dart-sass/issues/868. That led to me looking at
sass-embedded as a drop-in replacement for sass. I tried that, and yep,
it does what it says on the tin. Works the same, but async is way faster
and comparable to sync perf (with the benefit of much higher throughput
if you're doing a lot of stuff concurrently). My entire build wall-clock
time is now down to 8s, and the slowest css bundles are 4.5s.

Biggest downside seems to be that dart isn't compatible with musl libc,
so I had to swap the base docker image I was using for my esbuild
pipeline. Unclear to me if there are other tradeoffs.